feat: Add plugin that imports projects from a dir
This commit adds a plugin that checks if an environment variable
`MODEL_INBOX_DIRECTORIES` is set and points to an existent directory or
multiple directories (`:` separated). If so it checks if the specified
directories contains Eclipse projects (at least a `.project` file is
given) and loads them into the workspace. The plugin can be put into the
`dropins` folder of Capella and activates during startup. The plugin
logs it's actions and errors to Capella's error log.